+++ pkg/util/k8sutil/probes/probes_test.go
@@ -25,6 +25,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	core "k8s.io/api/core/v1"
+
+	"github.com/arangodb/kube-arangodb/pkg/apis/shared"
 )
 
 func TestCreate(t *testing.T) {
@@ -32,7 +34,7 @@ func TestCreate(t *testing.T) {
 	secret := "the secret"
 
 	// http
-	config := HTTPProbeConfig{path, false, secret, 0, 0, 0, 0, 0, 0}
+	config := HTTPProbeConfig{path, false, secret, "", 0, 0, 0, 0, 0}
 	probe := config.Create()
 
 	assert.Equal(t, probe.InitialDelaySeconds, int32(15*60))
@@ -44,17 +46,17 @@ func TestCreate(t *testing.T) {
 	assert.Equal(t, probe.Handler.HTTPGet.Path, path)
 	assert.Equal(t, probe.Handler.HTTPGet.HTTPHeaders[0].Name, "Authorization")
 	assert.Equal(t, probe.Handler.HTTPGet.HTTPHeaders[0].Value, secret)
-	assert.Equal(t, probe.Handler.HTTPGet.Port.IntValue(), 8529)
+	assert.Equal(t, probe.Handler.HTTPGet.Port.String(), shared.ServerPortName)
 	assert.Equal(t, probe.Handler.HTTPGet.Scheme, core.URISchemeHTTP)
 
 	// https
-	config = HTTPProbeConfig{path, true, secret, 0, 0, 0, 0, 0, 0}
+	config = HTTPProbeConfig{path, true, secret, "", 0, 0, 0, 0, 0}
 	probe = config.Create()
 
 	assert.Equal(t, probe.Handler.HTTPGet.Scheme, core.URISchemeHTTPS)
 
 	// http, custom timing
-	config = HTTPProbeConfig{path, false, secret, 0, 1, 2, 3, 4, 5}
+	config = HTTPProbeConfig{path, false, secret, "", 1, 2, 3, 4, 5}
 	probe = config.Create()
 
 	assert.Equal(t, probe.InitialDelaySeconds, int32(1))